api-gateway/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
PyJWT
httpx
python-dotenv
